refactor(richmond-buildings): split livable_apartment into foo.rs + foo/*
Move the apartment module off `mod.rs` to the crate’s preferred layout:
`livable_apartment.rs` plus `livable_apartment/{room,entry,layout,program,tests}.rs`,
matching `eating_area` / `rectangular_livable_area`.

## Agent Dialogue

- **Developer:** Prefer modern submodule layout — `livable_apartment.rs` + `livable_apartment/*`, not `mod.rs`.
- **Agent:** Relocated the parent module to `livable_apartment.rs`, kept children under the directory, deleted `mod.rs`; fixed trait imports; `livable_apartment` lib tests pass.
